mainjs breakdown into country segments
