+++ html/argentina.html
@@ -24,13 +24,13 @@
   <section class="mainContainer"> 
     <article class="subContainer">
       <article class="alandmarkobjects">
-        Mynameis
+        Argentina
       </article>
       <article class="acitizenobjects">
-        David
+        What is happening
       </article>
     </article>
   </section>
-  <script type="module" src="../scripts/main.js"></script>
+  <script type="module" src="../scripts/argentina/amain.js"></script>
 </body>
 </html>
