+++ html/bolivia.html
@@ -24,13 +24,13 @@
   <section class="mainContainer"> 
     <article class="subContainer">
       <article class="blandmarkobjects">
-        Mynameis
+We are not in Bolivia but we should be
       </article>
       <article class="bcitizenobjects">
         David
       </article>
     </article>
   </section>
-  <script type="module" src="../scripts/main.js"></script>
+  <script type="module" src="../scripts/bolivia/bmain.js"></script>
 </body>
 </html>
