+++ html/japan.html
@@ -24,13 +24,12 @@
   <section class="mainContainer"> 
     <article class="subContainer">
       <article class="jlandmarkobjects">
-        Mynameis
-      </article>
+But this time in Japan      </article>
       <article class="jcitizenobjects">
         David
       </article>
     </article>
   </section>
-  <script type="module" src="../scripts/main.js"></script>
+  <script type="module" src="../scripts/japan/jmain.js"></script>
 </body>
 </html>
