+++ html/scotland.html
@@ -24,13 +24,13 @@
   <section class="mainContainer"> 
     <article class="subContainer">
       <article class="slandmarkobjects">
-        Mynameis
+        Are you working
       </article>
       <article class="scitizenobjects">
-        David
+        Probably not
       </article>
     </article>
   </section>
-  <script type="module" src="../scripts/main.js"></script>
+  <script type="module" src="../scripts/scotland/smain.js"></script>
 </body>
 </html>
